requirements.txt
beautifulsoup4==4.6.3
pypiwin32==223
pyserial==3.4
pywin32==224
selenium==3.141.0
urllib3==1.24.1
